Fix future denotation error by enabling FutureDefs
The miniphase is doing a deep traversal of the tree, breaking the
classical assumptions that miniphases only check the visiting node.
Therefore, it's indeed transforming future trees transformed by future
miniphases. The compiler checks this doesn't happen if a flag is not
enabled. In this case, enabling such a phase in the context works.