Make optimization recursive and lift up assigns
Repeated expressions that depend one of each other were not optimized
before. The following algorithm is brute-force and a more smart but
rather complicated algorithm could be used to detect these natural
dependencies between optimized trees and replace them on the fly.

Now, assignations are lifted up one level (to avoid breaking the
correctness of programs with side effects). Therefore, we do not
introduce a block with an assignation followed by a reference, as before,
but we assign the result independently. In order to do that, we
introduce the notion of entrypoints, that are dummy `ValDef`s used to
mark concrete locations in which these assignations need to be. These
entrypoints are necessary because trees may not have symbols, and we
need symbols to identify those spots in the transformer. Using trees
would lead us to a **very** complicated algorithm because inner trees
are supposed to have changed.