Fix recursivity of algorithm and idempotency
Some idempotent trees like `isInstanceOf` were not correctly identified
as idempotent. These changes make them so.

The recursivity of the optimizations work perfectly now.

Once we find a branch whose else is Unit, we don't even bother in
inspecting it. In the future, we could inspect the expressions in the
then clause because it could host repeated expressions defined locally.