Handle boolean operators in If predicates
Operators like && and || need to optimize only the leftmost tree of the
condition because there's no guarantee that the trees on the right side
will be evaluated.

However, the XOR operator (^) has this guarantee and it will need to be
initialized in both cases, which is rather nice.

Add tests that prove that this feature is working as expected in
different cases, where the depth of the tree varies and the boolean
conditions are repeated in inner trees.