Add support for non-idempotent inner methods
The algorithm is now able to optimize inner methods that are not
idempotent. Those which are idempotent are neglected at this moment and
will be fixed in an uncoming commit.

The PRE optimization follows the inner methods and only optimizes them
the first time they appear. If they happen to be inside if expressions,
templates, lambdas or try catches that **are inside top level
expressions in a parameter position**, their optimization will be
disabled.