Identify calls of idempotent parameterless methods
Modify `IdempotencyInference` to consider calls to parameterless methods
idempotent iff the methods are idempotent.

Before, the idempotency inference was rejecting all the top-level
`Ident` nodes, because methods usually need `Apply` or `TypeApply`.
However, this is not the case for calls of parameterless methods.

Also insert `println`s in test cases to ensure that calls of inner
functions to be tested are not considered idempotent.